connectX/requirements.txt
absl-py==2.1.0
agent @ file:///Users/meinczinger/src/github/kaggle/connectX/agent
astunparse==1.6.3
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
Chessnut==0.4.1
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
game @ file:///Users/meinczinger/src/github/kaggle/connectX/game
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
gymnasium==0.29.0
h5py==3.12.1
huggingface-hub==0.26.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle-environments==1.16.9
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
logger @ file:///Users/meinczinger/src/github/kaggle/connectX/logger
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mcts @ file:///Users/meinczinger/src/github/kaggle/connectX/mcts
mdurl==0.1.2
ml-dtypes==0.3.2
model @ file:///Users/meinczinger/src/github/kaggle/connectX/model
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.0
pillow==11.0.0
protobuf==4.25.5
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scipy==1.14.1
Shimmy==1.3.0
six==1.17.0
stable-baselines3==2.1.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
